feat: add auth/sign-up page with sign up server action
Add auth/sign-up page with a form to get the name, email and password of
the user and a submit button to run a client side validation and then a
server action to sign up using better-auth email and password
authentication.

Add types.ts in the route directory to define the zod schema of the form
for client side validation. Add actions.ts file to define the server
action for signing up.

Add shadcn package along with button, card, form, input, label and
sonner components.