feat: add auth/sign-in page with sign in server action
Add auth/sign-in page with a form to get the email and password of the
user and a submit button to run a client side validation and then a
server action to sign in using better-auth email and password
authentication.

Add types.ts in the route directory to define the zod schema of the form
for client side validation. Add actions.ts file to define the server
action for signing in.