feat: add middleware to check if already signed in with cookies
Add a chained middleware logic in middlewares/index.ts and middleware.ts
and define a middleware withAuth to check if the request is in /auth/*
routes and if the user is already signed in, then the user is redirected
to the root page. The middleware sets a cookie with key 'toast'
containing the toast level and message in the response for displaying a
toast.

Add a component CookieToaster to read the 'toast' cookie on the first
render and then display a toast with the message and level. It then
deletes the cookie. Add the component in the root layout.tsx to check
for the cookie on the first render of each page.