feat: add withRole middleware for role based access control
Update drizzle schema to include a role column in the user table, where
'admin', 'author' and 'user' are permitted values.

Add middlewares/with-role middleware to protect the /admin/* routes. It
redirects to the auth/sign-in page if no one is signed in or redirects
to the root page if the user is not of admin or author role.