feat: add submit action to print the editor form values to the console
Refactor the TipTap component to RichTextEditor component and remove the
MenuBar from the RichTextEditor component to decouple the two
components.

Remove the content field from the zod type and add a description field.
Add a description FormItem in the EditorForm component.

Add a form submit action to print the form values along with the content
in the rich text editor to the console.