feat: add post table, action to update a post and ISR for editor/[slug]
Introduce a post table in the drizzle schema for storing the different
blog posts. Move the editor to a dynamic route editor/[slug] to edit
different posts with the slug.

Add a server action to update a post for a given slug and revalidate the
editor/[slug] page for the page edited.

Update the editor page and component to fetch the data of the post from
the database and show it in the editor.