feat: add image upload functionality with vercel blob and image uploader
Introduce @vercel/blob as an npm package. Modify drizzle schema to add
an image table to keep track of images uploaded by each user. Add an
ImageSelector card component to show an image selection window in a grid
and buttons to upload and delete images in the footer.

Add server actions to upload and delete images from vercel blob and
invoke the actions whenever the corresponding buttons are pressed.

Add EditorContext to provide states for whether the Image selector
window is opened and which image is selected.

Update next.config.ts to allow showing images from vercel blob storage
for the project.