feat: add global styles with button, input and textarea ui components
Update root layout.tsx file to include Archivo and Montserrat as the
base fonts for the app. Update globals.css file to define the color
palette and base styles for the header and body tags of the app.

Add Button, Input, InputErrorMessage, PasswordInput and TextArea UI
components. The input.tsx file contain the input block components for
the label and placeholder and the InputErrorMessage component. The
TextArea component extends the Input block components while the
PasswordInput component extends the Input component and adds a right
slot for the toggle button.