feat: add /auth pages with sign up and sign in actions with middleware
Introduce /auth/sign-in and /auth/sign-up pages with sign in and sign up
forms respectively. Add validated actions for sign in and sign up forms.

Add a middleware for /auth routes to redirect to the home page if a
better-auth session cookie is already present in the request headers.

Update globals.css file to add styles for anchor tags.