feat: add dialog ui component with backdrop styling
Add Dialog UI component to show modal windows and dialogs with a
useDialog hook that returns the RefObject to the dialog tag.

Update the root layout and global styles to wrap the whole content of
the app inside a div with body id to apply blurring and a black overlay
whenever a dialog is open to exclude the toast from getting blurred.