feat: add ABAC pattern and define user roles and permissions
Add the admin plugin for better-auth and define the post resource along
with its permissions. Refactor validated action pattern to rename
ValidatedActionResult to ActionResult.

Add a hasPermission() function in actions module in lib folder to check
if the user with the current session has a list of permissions.

Define an actions module in lib folder with a accessWithControl() to
perform Attribute Based Access Control.